requirement.txt
python-binance
flask
numpy
pandas